requirements.txt
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
asttokens==3.0.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
duckdb==0.9.2
et_xmlfile==2.0.0
executing==2.2.1
Faker==19.13.0
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.61.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.65.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.2
groq==1.0.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ipython==8.38.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.7.5
matplotlib-inline==0.2.1
numpy==1.26.4
openai==1.109.1
openpyxl==3.1.5
packaging==26.0
pandas==1.5.3
pandasai==2.0
parso==0.8.6
pillow==10.4.0
polars==1.38.1
polars-runtime-32==1.38.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scipy==1.10.1
seaborn==0.12.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.47
sqlglot==25.34.1
stack-data==0.6.3
tenacity==9.1.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.6
